Implemented a WIP-system for bot abilities. Rewritten program types to use dynamic IDs.
Implemented a database for program types.
Program types are now uniquely identified with a string typename, which dynamically resolves to an internal ID.
Implemented a factory template for unique IDs (UidFactory).
ProgramFactory now uses an UidFactory.
Updated simtest.cpp for the new changes.
Updated gamecontroller.h for the new changes.