Implemented a replay system for viewing arbitrary simulation timepoints.
Simplification of AbilityFactory and ProgramFactory by making them inherit DefaultUidFactory.
Fixed oversight/bug for PlayerData, where it was saved, but never loaded.
Minor addition to Save, to aid in debugging save/load savedata alignment.
Migrated some data from Simulation. Moved the statecopy to the replay manager, and the phase variables to State.
Added Simulation::clear to reset the simulation to a clean state.
Fixed severe bug in World::collide and DataBehaviourT::InternalDatabase::addTypeExt(). Serves me right for not -Walling myself in.
Added replay testing to simtest.cpp and updated it according to the new changes.