Implemented template functions for retrieval of common simulation and extsim components as per issue #2.
Most of the generic system stems from a few X macro defines that take advantage of redefinition of _SIM_X and _EXTS_X.
Added CommonTemplate.cpp/.h in both extsim/ and simulation/ that implements convenience/helper functions.
Simplified part of DefaultUidFactory as a result of the changes, and related factory implementations.
Standardized a few names, making them more consistent.
Simplified ExtS::IdList as a result of the changes.