Removed Sim::Input, as it is obsolete due to improvements to the InputBarrier and ReplayManager.
Altered affected TypeRule implementations for the new changes.
Added a function to simulation factories to create objects directly.
Switched to manual ID allocation for Sim::State, still need to remove auto-allocation.

Comment: Input is now passed from TypeRule objects directly to Sim::State.
Input may be passed directly to the simulation as if using Sim::Simulation
without ExtSim, however this is not picked up by the ReplayManager unless
a full simulation copy is created. This makes direct input on the first
phase, which is always a full state copy, possible.